FIX: correctly order resource/datum generation documents
In the RunEngine collect_asset_docs is called before collect thus
there was an off-by-one bug where the assetregistry documents
collected at any step were from the previous step.  This means we
violated some invariants (always see the asset docs before the event
that uses it) and the last set of asset docs never got emitted (or
might as part of the next fly scan!!).